requirements.txt
# Core FastAPI dependencies
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.27.0,<0.28.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# Optional, but often useful
python-multipart>=0.0.9,<0.0.10 # For file uploads
Jinja2>=3.1.0,<3.2.0 # For HTML templating (if needed)
